finish all exercises

+++ exercises/conversions/using_as.rs
@@ -6,11 +6,10 @@
 // and returns the proper type.
 // Execute `rustlings hint using_as` or use the `hint` watch subcommand for a hint.
 
-// I AM NOT DONE
 
 fn average(values: &[f64]) -> f64 {
     let total = values.iter().sum::<f64>();
-    total / values.len()
+    total / values.len() as f64
 }
 
 fn main() {
@@ -24,6 +23,6 @@ mod tests {
 
     #[test]
     fn returns_proper_type_and_value() {
-        assert_eq!(average(&[3.5, 0.3, 13.0, 11.7]), 7.125);
+        assert_eq!(average(&[3.5 as f64, 0.3 as f64, 13.0 as f64, 11.7 as f64]), 7.125);
     }
 }
